pyproject.toml
[project]
name = "agentic-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.68.0",
    "langchain>=1.2.13",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
